+++ common/Cargo.toml
@@ -1,29 +1,24 @@
 [package]
-name = "identity"
+name = "common"
 version = "0.1.0"
 authors = ["Master Union <masteruniondoo@gmail.com>"]
 edition = "2021"
 
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
 [dependencies]
 ink = { version = "4.2.0", default-features = false }
-common = { path = "../../common", default-features = false }
-
 scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
 scale-info = { version = "2.3", default-features = false, features = ["derive"], optional = true }
 
-[dev-dependencies]
-ink_e2e = "4.2.0"
-
 [lib]
-path = "lib.rs"
+path = "src/lib.rs"
+crate-type = ["rlib"]
 
 [features]
 default = ["std"]
 std = [
     "ink/std",
-    "common/std",
     "scale/std",
     "scale-info/std",
 ]
-ink-as-dependency = []
-e2e-tests = []
