+++ common/src/lib.rs
@@ -1,0 +1,4 @@
+//! Types used in multiple contracts.
+#![cfg_attr(not(feature = "std"), no_std)]
+
+pub mod types;
