+++ contracts/address-book/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 
 [dependencies]
 ink = { version = "4.2.0", default-features = false }
+common = { path = "../../common", default-features = false }
 
 scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
 scale-info = { version = "2.6", default-features = false, features = ["derive"], optional = true }
@@ -21,6 +22,7 @@ path = "lib.rs"
 default = ["std"]
 std = [
     "ink/std",
+    "common/std",
     "scale/std",
     "scale-info/std",
     "identity/std",
